Rating passé en Double au lieu de Integer + ajout des différents boutons de fiche-film.xhtml

+++ HomeCinema-lib/src/dtos/FilmDto.java
@@ -20,5 +20,5 @@ public class FilmDto implements Serializable {
     public Date release_date;
     public String country;
     public Integer runtime;
-    public Integer rating;
+    public Double rating;
 }
